requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.1.5
django-allauth==65.3.1
django-crispy-forms==2.3
django-filter==24.3
djangorestframework==3.15.2
filelock==3.17.0
numpy==2.2.2
pandas==2.2.3
platformdirs==4.3.6
python-dateutil==2.9.0.post0
pytz==2025.1
RapidFuzz==3.12.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
virtualenv==20.29.1
